+++ crates/cgp-component/src/types/delegate_to.rs
@@ -1,3 +1,0 @@
-use core::marker::PhantomData;
-
-pub struct DelegateTo<Components>(pub PhantomData<Components>);