+++ crates/cgp-field/src/traits/has_field.rs
@@ -1,6 +1,8 @@
 use core::marker::PhantomData;
 use core::ops::Deref;
 
+use cgp_component::UseContext;
+
 pub trait HasField<Tag> {
     type Field;
 
@@ -22,5 +24,16 @@ where
 
     fn get_field(&self, tag: PhantomData<Tag>) -> &Self::Field {
         self.deref().get_field(tag)
+    }
+}
+
+impl<Context, Tag, Field> FieldGetter<Context, Tag> for UseContext
+where
+    Context: HasField<Tag, Field = Field>,
+{
+    type Field = Field;
+
+    fn get_field(context: &Context, _tag: PhantomData<Tag>) -> &Self::Field {
+        context.get_field(PhantomData)
     }
 }