+++ crates/cgp-type/src/traits/has_type.rs
@@ -1,4 +1,4 @@
-use cgp_component::{derive_component, DelegateComponent, DelegateTo, HasComponents, UseContext};
+use cgp_component::{derive_component, DelegateComponent, HasComponents, UseContext, UseDelegate};
 
 #[derive_component(TypeComponent, ProvideType<Context>)]
 pub trait HasType<Tag> {
@@ -12,7 +12,7 @@ where
     type Type = Context::Type;
 }
 
-impl<Context, Tag, Components, Type> ProvideType<Context, Tag> for DelegateTo<Components>
+impl<Context, Tag, Components, Type> ProvideType<Context, Tag> for UseDelegate<Components>
 where
     Components: DelegateComponent<Tag>,
     Components::Delegate: ProvideType<Context, Tag, Type = Type>,